layers: Add a Image Layout CoreCheck file
